gcs_mcp_server/requirements.txt
fastmcp
google-cloud-storage
uvicorn